Use int32_t instead of 64 bit for OUT_FENCE_PTR.
OUT_FENCE_PTR was changed to int32_t with this patch:
https://patchwork.kernel.org/patch/9515719/

Jira: None.
Tests: No regressions on Linux and Android.

+++ common/compositor/compositor.cpp
@@ -154,7 +154,7 @@ bool Compositor::DrawOffscreen(std::vector<OverlayLayer> &layers,
   return true;
 }
 
-void Compositor::InsertFence(uint64_t fence) {
+void Compositor::InsertFence(uint32_t fence) {
   renderer_->InsertFence(fence);
 }
